+++ lib/modern_player.dart
@@ -2,14 +2,11 @@ library modern_player;
 
 export 'src/modern_player.dart' show ModernPlayer;
 export 'src/modern_players_enums.dart'
-    show
-        ModernPlayerSourceType,
-        ModernPlayerSubtitleSourceType,
-        ModernPlayerAudioSourceType;
+    show ModernPlayerSubtitleSourceType, ModernPlayerAudioSourceType;
 export 'src/modern_player_options.dart'
     show
         ModernPlayerOptions,
-        ModernPlayerQualityOptions,
+        ModernPlayerVideoData,
         ModernPlayerControlsOptions,
         ModernPlayerThemeOptions,
         ModernPlayerProgressSliderTheme,